more restructuring and data binding
