+++ .gitignore
@@ -9,8 +9,8 @@ platforms/
 !webpack.config.js
 
 # Some Sass-compiled files
-app/app.android.css
-app/app.ios.css
+/app/app.android.css
+/app/app.ios.css
 app/views/app-root/app-root.android.css
 app/views/app-root/app-root.ios.css
