+++ app/views/login/login-page.xml
@@ -1,13 +1,12 @@
 <Page actionBarHidden="true" loaded="loaded">
-    <!--<ActionBar title="Assetti"></ActionBar>-->
     <FlexboxLayout>
         <StackLayout class="form">
-            <Image src="~/images/assetti.jpg" />
-            <Label class="h1" text="Assetti Mobile" />
-            <TextField id="email" text="{{ userData.email }}" hint="Email" keyboardType="email" autocorrect="false" autocapitalizationType="none" class="input input-border" />
-            <TextField secure="true" text="{{ userData.password }}" hint="Password" class="input input-border" />
-            <Button text="Sign in" class="btn btn-primary" tap="onTap" />
-            <Button text="By Pass Login // TEST MODE" class="btn btn-primary" tap="onTapByPass" />
+            <Image src="~/images/assetti.jpg"></Image>
+            <Label class="h1" text="Assetti Mobile"></Label>
+            <TextField id="email" text="{{ userData.email }}" hint="Email" keyboardType="email" autocorrect="false" autocapitalizationType="none" class="input input-border"></TextField>
+            <TextField secure="true" text="{{ userData.password }}" hint="Password" class="input input-border"></TextField>
+            <Button text="Sign-in" class="btn btn-primary" tap="tapLogin"></Button>
+            <Button text="Sign-in Bypass" class="btn btn-primary" tap="tapLoginBypass"></Button>
         </StackLayout>
     </FlexboxLayout>
 </Page>
