+++ package.json
@@ -17,23 +17,21 @@
   "dependencies": {
     "base-64": "^0.1.0",
     "jsrsasign": "^8.0.12",
+    "nativescript": "^5.3.0",
     "nativescript-background-http": "^3.4.0",
-    "nativescript-camera": "^4.2.0",
+    "nativescript-camera": "^4.4.0",
     "nativescript-imagepicker": "^6.1.2",
-    "nativescript-theme-core": "~1.0.4",
+    "nativescript-theme-core": "^1.0.4",
     "nativescript-ui-autocomplete": "^4.0.0",
-    "nativescript-ui-chart": "^3.11.1",
-    "nativescript-ui-listview": "^6.0.0",
-    "nativescript-ui-sidedrawer": "~5.1.0",
-    "rxjs": "~6.3.0",
-    "tns-core-modules": "~5.2.0",
-    "typescript": "^3.3.4000"
+    "nativescript-ui-listview": "^6.1.0",
+    "nativescript-ui-sidedrawer": "^6.0.0",
+    "tns-core-modules": "^5.3.0",
+    "typescript": "^3.4.1"
   },
   "devDependencies": {
-    "nativescript-dev-sass": "~1.6.0",
-    "nativescript-dev-typescript": "~0.8.0",
-    "nativescript-dev-webpack": "~0.20.0",
-    "tslint": "~5.11.0"
+    "nativescript-dev-sass": "^1.7.0",
+    "nativescript-dev-typescript": "^0.9.0",
+    "nativescript-dev-webpack": "^0.21.0"
   },
   "gitHead": "f548ec926e75201ab1b7c4a3a7ceefe7a4db15af",
   "readme": "Assetti NativeScript Application"
